Split classes from core/metrics_manager.js into multiple files
