+++ core/flyout_metrics_manager.js
@@ -1,0 +1,99 @@
+/**
+ * @license
+ * Copyright 2021 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * @fileoverview Calculates and reports flyout workspace metrics.
+ * @author aschmiedt@google.com (Abby Schmiedt)
+ */
+'use strict';
+
+goog.provide('Blockly.FlyoutMetricsManager');
+
+goog.require('Blockly.MetricsManager');
+goog.require('Blockly.utils.object');
+
+goog.requireType('Blockly.IFlyout');
+goog.requireType('Blockly.WorkspaceSvg');
+
+
+/**
+ * Calculates metrics for a flyout's workspace.
+ * The metrics are mainly used to size scrollbars for the flyout.
+ * @param {!Blockly.WorkspaceSvg} workspace The flyout's workspace.
+ * @param {!Blockly.IFlyout} flyout The flyout.
+ * @extends {Blockly.MetricsManager}
+ * @constructor
+ */
+Blockly.FlyoutMetricsManager = function(workspace, flyout) {
+  /**
+   * The flyout that owns the workspace to calculate metrics for.
+   * @type {!Blockly.IFlyout}
+   * @protected
+   */
+  this.flyout_ = flyout;
+
+  Blockly.FlyoutMetricsManager.superClass_.constructor.call(this, workspace);
+};
+Blockly.utils.object.inherits(
+    Blockly.FlyoutMetricsManager, Blockly.MetricsManager);
+
+/**
+ * Gets the bounding box of the blocks on the flyout's workspace.
+ * This is in workspace coordinates.
+ * @return {!SVGRect|{height: number, y: number, width: number, x: number}} The
+ *     bounding box of the blocks on the workspace.
+ * @private
+ */
+Blockly.FlyoutMetricsManager.prototype.getBoundingBox_ = function() {
+  try {
+    var blockBoundingBox = this.workspace_.getCanvas().getBBox();
+  } catch (e) {
+    // Firefox has trouble with hidden elements (Bug 528969).
+    // 2021 Update: It looks like this was fixed around Firefox 77 released in
+    // 2020.
+    var blockBoundingBox = {height: 0, y: 0, width: 0, x: 0};
+  }
+  return blockBoundingBox;
+};
+
+/**
+ * @override
+ */
+Blockly.FlyoutMetricsManager.prototype.getContentMetrics = function(
+    opt_getWorkspaceCoordinates) {
+  // The bounding box is in workspace coordinates.
+  var blockBoundingBox = this.getBoundingBox_();
+  var scale = opt_getWorkspaceCoordinates ? 1 : this.workspace_.scale;
+
+  return {
+    height: blockBoundingBox.height * scale,
+    width: blockBoundingBox.width * scale,
+    top: blockBoundingBox.y * scale,
+    left: blockBoundingBox.x * scale,
+  };
+};
+
+/**
+ * @override
+ */
+Blockly.FlyoutMetricsManager.prototype.getScrollMetrics = function(
+    opt_getWorkspaceCoordinates, opt_viewMetrics, opt_contentMetrics) {
+  var contentMetrics = opt_contentMetrics || this.getContentMetrics();
+  var margin = this.flyout_.MARGIN * this.workspace_.scale;
+  var scale = opt_getWorkspaceCoordinates ? this.workspace_.scale : 1;
+
+  // The left padding isn't just the margin. Some blocks are also offset by
+  // tabWidth so that value and statement blocks line up.
+  // The contentMetrics.left value is equivalent to the variable left padding.
+  var leftPadding = contentMetrics.left;
+
+  return {
+    height: (contentMetrics.height + 2 * margin) / scale,
+    width: (contentMetrics.width + leftPadding + margin) / scale,
+    top: 0,
+    left: 0,
+  };
+};
